Extract method interceptors from ShadowWranger.
